refactor: inline vocabulary search shortcut in page component
Remove single-use useSearchShortcut hook per review feedback.
